MLPipeline/pyproject.toml
[tool.poetry]
name = "mlpipelinelecture"
version = "0.1.0"
description = ""
authors = ["Cezar Sas <cezar.sas@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
pytorch-lightning = "^1.5.10"
torch = "^1.11.0"
torchvision = "^0.12.0"
matplotlib = "^3.5.1"
torchmetrics = "^0.7.2"
tensorflow = "^2.8.0"
hydra-core = "^1.1.1"
rich = "^12.0.1"
hydra-colorlog = "^1.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"